Filter resources based on search query
Ie if we add

    COLLECTION_FILTERS = {
      inactive: ->(resources) { resources.where("login_at < ?", 1.week.ago) }
    }

to a dashboard, we can query the resources of that dashboard with

    bob inactive:

to find users named "bob" who hasn't logged in the last week.

If you already had the `inactive` scope you could define the filter like so to
take advantage of existing ActiveRecord scopes (and other class methods on the
resource class).

    COLLECTION_FILTERS = {
      inactive: ->(resources) { resources.inactive }
    }

While the chosen hash-based syntax is a bit more verbose than simply exposing
already defined scopes like so:

    # app/dashboards/customer_dashboard.rb
    COLLECTION_FILTERS = [:inactive]

it allows us to define filters for use in Administrate without having to clutter
the resource classes with scopes.

It still allows us to add the simpler syntax in a backwards compatible way at
some point down the line if we feel the need. For example it could end up
looking like:

    COLLECTION_FILTERS = {
      vip: :vip, # This could call the method `vip` on resources
      inactive: ->(resources) { resources.where("login_at < ?", 1.week.ago) }
    }

but that's a step I am not ready to take just yet.

It also opens up for a way to have filters that take arguments, ie we could
define:

    COLLECTION_FILTERS = {
      before: ->(resources, args) {
                resources.where("created_at < ?", args)
              }
    }

and query it as

    before:2016-01-12

which, again, is a step I am not ready to take here and now.